Move a bunch of Bridge classes to the namespace
