+++ src/Tgstation.Server.Host/Components/IInstanceManager.cs
@@ -1,6 +1,6 @@
 ﻿using System.Threading;
 using System.Threading.Tasks;
-using Tgstation.Server.Host.Components.Interop;
+using Tgstation.Server.Host.Components.Interop.Bridge;
 using Tgstation.Server.Host.Models;
 
 namespace Tgstation.Server.Host.Components
