+++ src/Tgstation.Server.Host/Components/Interop/Bridge/IBridgeRegistration.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace Tgstation.Server.Host.Components.Interop
+namespace Tgstation.Server.Host.Components.Interop.Bridge
 {
 	/// <summary>
 	/// Represents a registration of an interop session.
